backend/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Config & validation
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP
python-multipart>=0.0.9